Refactor Monolithic generation.py into submodule
3799 line fine refactored into a 5 file submodule.
