+++ TrailPrint3D/utils/generation/__init__.py
@@ -1,0 +1,5 @@
+"""Map/trail generation pipeline package.
+
+Split from the former monolithic ``generation.py`` module into phase-grouped
+submodules.
+"""
